pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jippity-block-tv"
version = "0.1"
requires-python = ">= 3.10"
dependencies = [
  "requests",
  "python-dotenv",
  "langchain",
  "langchain-openai",
  "validators"
]